Allow writing custom `riff::ChunkContents` to WAV
Include a new `write_custom` method allowing to include custom
`riff::ChunkContents` between the `fmt ` and the `data` chunks of the
WAV file.

In order to DRY move all the logic from `write` to `write_custom` so
that `write` calls `write_custom` with `custom_chunk_contents = None`.
This bumps the semver to 1.1.0.